Added open and save features to the editor, with huge refactoring
This commit first brings the save and open files functionnalities to
the editor. We also deleted some buttons in the main grid and removed
some spacing.
This commit also comes with a lot of refactoring like a own class
for the header bar and a own class for the editor, with the goal to be
easily able to make multiple instances of the editor. This refactoring
is made to avoid huge pain in the ass later on, when we will need to
add new functionnalities to TeXSharp.