Creation of a new branch for controversial zoom functionnality
Adding a zoom functionnaly creates two different cases for now :
1- We use Picture widget which make the image fill the available space
but the zoom doesn't work because Picture can't be resized, or not
easily (for now). This is default behavior on 'main' branch
2- We use Image widget which is easy to resize, so the zoom work. But
Image doesn't fill like Picture which makes weird gaps appearing. This
change is not very pleasing so it's added to the 'zoom' branch.