Patched dirty compile option, with PDF viewer updating, and cleanup
The compile option was compiling in TeXSharp source folder, not in the
tex file directory, this was fixed. The file only compile if the file
currently in the editor is saved, and saves the file before compiling.
On the startup of TeXSharp, the PDF viewer is hidden. When you open a
file, it looks for a corresponding pdf and opens it. After compiling,
the PDF viewer automatically refresh with the new PDF.
Added menu button for LaTeX in the menu bar and added sub button to compile.
For the cleanup: a commented chunk of code was removed in order to
keep the code base clean.