Added possibility to create a new file when editing a file
What this means is that it is now possible to create a new file, even
when editing an already existing one without having to close the app
and reopen it to get a new file. Next step it to simply add the new file
to a new tab, instead of replacing the current one.